Folder structure formatted and tested
