+++ .gitignore
@@ -4,7 +4,6 @@
 tree.py
 *testTlink.py
 consoleBot.py
-knowledge_base.csv
 alfred/
 .env
 test
