+++ app/services/ollama.py
